refactor(starterkit): split the format and time converters into one type per file
- TimeSpanFormatConverter now reports every unusable format, not only the
  first: the once-only flag hid every later failure.
- NumberFormatConverters and TimeConverters become one file per type, next to
  CultureInfoMode, with the usings trimmed to what each file needs.
- Enum member summaries move to the multiline form the codebase uses
  everywhere else.
