+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/AbbreviatedNumberConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 90f41b8e51e94121b2fad65a29c391dc
