+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/ByteSizeConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c562c49d412143ae9a8345e232801664
