+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/CurrencyConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 51413a32e13a4dcfb603d8b5b9ba9c31
