+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/OrdinalConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fb0ffd562db54f32ac63809d0bb84782
