+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/PluralRule.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e71e531e1fa0477d826c0e4ad91aaf04
