+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/PluralizeConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ad42a4e54f3c4074bebe2f21e831fbf7
