+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/RomanNumeralConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 27127c50e57f4534b44633870baab17d
