+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/SignedNumberStringConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e7ff0524f8fc45edbd069bb217dbab9f
