+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/SymbolPosition.cs
@@ -1,0 +1,19 @@
+// ReSharper disable once CheckNamespace
+namespace Aspid.MVVM.StarterKit
+{
+    /// <summary>
+    /// Where <see cref="CurrencyConverter"/> puts the symbol.
+    /// </summary>
+    public enum SymbolPosition
+    {
+        /// <summary>
+        /// Before the number.
+        /// </summary>
+        Before,
+
+        /// <summary>
+        /// After the number.
+        /// </summary>
+        After,
+    }
+}
