+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Strings/SymbolPosition.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fd2804551ec240d98d3004b0bc07efbc
