+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Times/DateTimeToBoolConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 49ac18a4cdf14390ab964789ff1c4d02
