+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Times/SecondsToTimeSpanConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 716c0c00c1364505ba1cb0f0a09f807c
