+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Times/TimeSpanFormatConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5ffb7acb25974955bdb247dda0f87a9c
