+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Times/TimeSpanToNumberConverter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 96c8c4c9957c4d25a8e7f89cc042e492
